# api/app_openapi/types.gen.go
// Package app_openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.8.2 DO NOT EDIT.
package app_openapi

import (
	"time"
)

// Link defines model for Link.
type Link struct {
	CreatedAt time.Time  `json:"createdAt"`
	DeletedAt *time.Time `json:"deletedAt,omitempty"`
	ExpiredAt *time.Time `json:"expiredAt,omitempty"`
	Hits      int32      `json:"hits"`
	TargetUrl string     `json:"targetUrl"`
	Token     string     `json:"token"`
}

// RequestShortUrl defines model for RequestShortUrl.
type RequestShortUrl struct {
	ExpiredInDays *int32 `json:"expiredInDays,omitempty"`
	TargetUrl     string `json:"targetUrl"`
}

// ResponseShortUrl defines model for ResponseShortUrl.
type ResponseShortUrl struct {
	ShortUrl     string  `json:"shortUrl"`
	ShortUrlInfo *string `json:"shortUrlInfo,omitempty"`
}

// CreateShortUrlJSONBody defines parameters for CreateShortUrl.
type CreateShortUrlJSONBody RequestShortUrl

// CreateShortUrlJSONRequestBody defines body for CreateShortUrl for application/json ContentType.
type CreateShortUrlJSONRequestBody CreateShortUrlJSONBody